refactor so models and modules loaded only when necessary
