+++ tests/test_chat_interact.py
@@ -5,19 +5,21 @@ import pytest
 from typer.testing import CliRunner
 
 from reginald.cli import cli
-from reginald.models.models.hello import Hello
-from reginald.run import run_chat_interact
+from reginald.models.chat_interact import art, run_chat_interact
+from reginald.models.simple.hello import Hello
 
 runner = CliRunner()
+art_split = art.splitlines()
 
 
 def test_chat_cli():
     """Test sending an input `str` via `cli` and then exiting."""
     result = runner.invoke(cli, ["chat"], input="What's up dock?\nexit\n")
     term_stdout_lines: list[str] = result.stdout.split("\n")
-    assert term_stdout_lines[0] == ">>> "
-    assert term_stdout_lines[1] == "Reginald: Hello! How are you?"
-    assert term_stdout_lines[2] == ">>> "
+    assert term_stdout_lines[: len(art_split)] == art_split
+    assert term_stdout_lines[len(art_split) + 1] == ">>> "
+    assert term_stdout_lines[len(art_split) + 2] == "Reginald: Hello! How are you?"
+    assert term_stdout_lines[len(art_split) + 3] == ">>> "
 
 
 def test_chat_cli_no_stream():
@@ -26,9 +28,13 @@ def test_chat_cli_no_stream():
         cli, ["chat", "--no-streaming"], input="What's up dock?\nexit\n"
     )
     term_stdout_lines: list[str] = result.stdout.split("\n")
-    assert term_stdout_lines[0] == ">>> "
-    assert term_stdout_lines[1] == "Reginald: Let's discuss this in a channel!"
-    assert term_stdout_lines[2] == ">>> "
+    assert term_stdout_lines[: len(art_split)] == art_split
+    assert term_stdout_lines[len(art_split) + 1] == ">>> "
+    assert (
+        term_stdout_lines[len(art_split) + 2]
+        == "Reginald: Let's discuss this in a channel!"
+    )
+    assert term_stdout_lines[len(art_split) + 3] == ">>> "
 
 
 def test_chat_interact_exit():
@@ -58,14 +64,16 @@ def test_chat_interact_bye():
 def test_chat_interact_clear_history():
     result = runner.invoke(cli, ["chat"], input="clear_history\n")
     term_stdout_lines: list[str] = result.stdout.split("\n")
-    assert term_stdout_lines[0] == ">>> "
-    assert term_stdout_lines[1] == "Reginald: No history to clear."
-    assert term_stdout_lines[2] == ">>> "
+    assert term_stdout_lines[: len(art_split)] == art_split
+    assert term_stdout_lines[len(art_split) + 1] == ">>> "
+    assert term_stdout_lines[len(art_split) + 2] == "Reginald: No history to clear."
+    assert term_stdout_lines[len(art_split) + 3] == ">>> "
 
 
 def test_chat_interact_slash_clear_history():
     result = runner.invoke(cli, ["chat"], input="\clear_history\n")
     term_stdout_lines: list[str] = result.stdout.split("\n")
-    assert term_stdout_lines[0] == ">>> "
-    assert term_stdout_lines[1] == "Reginald: No history to clear."
-    assert term_stdout_lines[2] == ">>> "
+    assert term_stdout_lines[: len(art_split)] == art_split
+    assert term_stdout_lines[len(art_split) + 1] == ">>> "
+    assert term_stdout_lines[len(art_split) + 2] == "Reginald: No history to clear."
+    assert term_stdout_lines[len(art_split) + 3] == ">>> "
